20210327 : add main, func_dt_handle, func_plot ; rewrite func_dt_handle
