--- data/README.md
@@ -1,7 +1,0 @@
-# NDU_Sunsing_Lidar
-This is for analyzing the lidar from National Defense University which will be used in the course of NTU-AS 2021 Field Measurement of Atmospheric Environment II
-
-[NDU Lidar SOP hackmd](https://hackmd.io/rlmIsb0UQCqlHdhiee7wYg)
-
-
-
